stockapi/requirements.txt
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
Flask==1.0.2
Flask-Cors==3.0.10
Flask-Login==0.4.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
hvac==0.11.2
idna==2.8
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
marshmallow==3.14.0
multitasking==0.0.9
oauthlib==3.0.1
psycopg2==2.9.2
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.21.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.24.3
Werkzeug==2.0.2
setuptools~=49.2.1
beautifulsoup4==4.10.0
soupsieve==2.3.1
